+++ pkg/config/loader.go
@@ -8,13 +8,6 @@ import (
 	"gopkg.in/yaml.v3"
 )
 
-const (
-	// The address where all configuration files are stored.
-	//
-	// TODO: Better for user to specify this as a parameter in a function not us hardcoding it
-	address string = "build/config/"
-)
-
 var (
 	// Main project configs
 	cfg                     interface{}
@@ -32,20 +25,11 @@ func Parse(path string, cfg interface{}, errorOnFileNotFound bool) error {
 	}
 }
 
-// Reads "address_to_config_folder/config.yaml" file and save them in `cfg` variable
-func ReadProjectConfigs(pwd string, cfg interface{}) error {
-	if err := Parse(filepath.Join(pwd, address, "config.yaml"), cfg, true); err != nil {
-		return err
-	}
-	return nil
-}
-
 // Reads "env.yaml" config file(if exists) and save them in `cfg` variable
-func ReadLocalConfigs(pwd string, cfg interface{}) error {
-	if err := Parse(filepath.Join(pwd, "env.yaml"), cfg, false); err != nil {
+func ReadLocalConfigs(cfg interface{}) error {
+	if err := Parse("env.yaml", cfg, false); err != nil {
 		return err
-	}
-	if err := Parse(filepath.Join(pwd, "env.yml"), cfg, false); err != nil {
+	} else if err := Parse("env.yml", cfg, false); err != nil {
 		return err
 	}
 	return nil