+++ pkg/logging/logger.go
@@ -34,22 +34,15 @@ type LogBundle struct {
 // Takes options needed for logs configs and returns
 // a *Logger
 //
-// If no address in `opt.Path` is provided, for linux users:
-// "/var/log/project" and for windows users "c:\\logs\project"
-// address will be used as default path
+// If no address in `opt.Path` is provided,
+// "/var/log/project" address will be used as default
 func New(opt *Option) (Logger, error) {
 	if opt == nil {
 		return nil, errors.New("option can not be nil")
 	}
 
 	if opt.Path == "" {
-		if runtime.GOOS == "linux" {
-			opt.Path = "/var/log/project"
-		} else if runtime.GOOS == "windows" {
-			opt.Path = "c:\\\\logs\\project"
-		} else {
-			return nil, errOperationSystemNotSupported
-		}
+		opt.Path = "/var/log/project"
 	}
 	err := createAddress(opt.Path)
 	if err != nil {